+++ depthai-ros/CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package depthai-ros
 2.6.3 (2023-02-10)
 ___________
 * Camera calibration updates
+* Option to connect to the device via USB port id
 
 2.6.2 (2023-02-01)
 ___________